requirements.txt
fastapi
alembic
asyncpg
sqlalchemy[asyncio]
pydantic-settings
python-multipart
pyjwt[crypto]
uvicorn[standard]
bcrypt
torch